services/webapi/requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.32.1
pydantic==2.12.0
pydantic-settings==2.11.0
SQLAlchemy==2.0.43
alembic==1.16.5
psycopg2-binary==2.9.10
redis==5.3.1
httpx==0.28.1
python-multipart==0.0.12
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1
email-validator==2.2.0

